Fixed BacktraceCleaner to never return an empty backtrace.
Empty backtraces means you didn't run any code, which isn't the case,
and goes against the contract that Minitest.backtrace_cleaner expects.
This fixes a bug I've seen in a number of reports.

It would be nice if this got backported to whatever versions are
active, as this keeps coming back on minitest issues.