Fixes against the previous commit.
Happy to squash this before merge. Just separate so I can whittle
these down.

Most of the problem seems to be centered around the backtrace filter
being both for general use and for minitest... and I'm only aware of /
addressing minitest issues.

If they really need to be separate, then we can make a separate one
for minitest.