Add `render_in` support to `capture` helper
Currently when using some form helpers along with `render_in` the output
is rendered out of order due to the use of `capture`. This happens
because the `capture` helper calls `with_output_buffer` on `self` when
the passed in block might be called in a different rendering context
with its own `@output_buffer`. Since the block is being evaluated in
that different rendering context setting `@output_buffer` via
`with_output_buffer` doesn't actually capture the content, but instead
renders the content directly into the block's rendering context's
`output_buffer`.

For example, if we have the following template:

```erb
<%= form_for @person do |f| %>
  <%= render InputComponent.new(form: f) %>
<% end %>
```

and this very simplified component template:

```erb
<%# InputComponent template %>
<%= form.label do %>
  Name
  <%= form.text_field :name %>
<% end %>
```

We will get rendered HTML that looks something like:

```html
<form>
  Name
  <input type="text" name="name">
  <label for="name"><label>
</form>
```

When we would expect:

```html
<form>
  <label for="name">
    Name
    <input type="text" name="name">
  <label>
</form>
```

This happens because the `form_for` builder keeps a reference to the
template object it was instantiated which is the object that label calls
`capture` on.

To resolve this issue, each time `capture` is called we retrieve the
`block`s rendering context and check if it responds to `render_in` and
`with_output_buffer`. This lets us override the correct `@output_buffer`
so we can capture the content and render it in the correct part of the
template.