Don't recommend disabling CSRF protection [ci skip]
The previous example disables CSRF protection for JSON format requests,
which allows any request to bypass it by adding the right Accept header.

Accept is one of the few headers that doesn't trigger a CORS preflight:
https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#Examples_of_access_control_scenarios

Instead of disabling it entirely, the simple case where authentication
is handled without a session (e.g. when using API keys) can be solved by
switching to the `:null_session` strategy.

For more complex cases (e.g. an API that uses session authentication),
I don't think a copy-and-paste-able example is appropriate; the user
needs to understand CSRF protection in order to make an informed choice.