Fix a test that started failing due to the previous commit
test_calling_update_changing_ids_doesnt_change_counter_cache (in
activerecord/test/cases/associations/has_many_associations_test.rb)
started failing due to the previous change, because the target wasn't
updated before updating the counter_cache on the target (i.e. in
the `after_update` callback in `ActiveRecord::Associations::Builder.add_counter_cache_callbacks`,
`increment_counters` was being called on the old target, not the new one).

It seems to me that this was a bug lurking in the code that was
exposed by the previous commit, not a new bug inroduced by the
previous commit.