Don't show backtrace if a Rails command isn't found
If a Rake command isn't found, Rake would generate an error with a
backtrace. While this can be useful if you create your own Rake apps, it
makes no sense for the Rails command line which developers only consume.

Instead we can suggests commands, from the Rails and Rake namespaces,
and leave out the backtrace.