Generators should raise an error if an attribute has an invalid type
Generators can create invalid migrations when passing an invalid
attribute type. For example, when mixing up the name and type:

    bin/rails g model post string:title

This will generate an attribute for post with a column named `string`
of the type `title`, instead of a column named `title` of the type
`string`. Running the migration will result in an error as the type
`title` is not known to the database.

Instead of generating invalid files, the generator should raise an error
if the type is invalid. We validate the type by checking if it's a
default migration types. If the type isn't a default type, we can ask the
database connection if the type is valid. This uses the `valid_type?`
method defined on each database adapter, which returns true if the
adapter supports the column type.