Add write_fragment_and_content_for method
It didn't seem clean to make write_fragment also responsible for writing the content_for data, so
extract write_fragment_and_content_for that knows how to combine both fragment and content_for into
a single hash to write to the cache.

This also causes it to be consistent and always store as {_fragment: fragment} even when we call
write_fragment with a string.