Record all calls (args passed) to content_for
We need to know whether flush: true was passed to each content_for call so that when we restore the
fragment from cache, we can "reply" the content_for calls so that they content_for data ends up
being the same when reading from the cache as it was after the initial page load.

Because we need to record all args, we have to store the calls in an array instead of a hash.