Deprecate YAML loading from legacy format older than Rails 5.0
`LegacyYamlAdapter` was added at afc124c (and 4e72170), to be able to
load the previous version formatted YAML.

Now, that has became quite useless, since we sometimes removed/added
dedicated type classes which are referred in the legacy format (e.g.
`MysqlDateTime`, `MysqlString`, `SQLite3Integer`), so it is unlikely
that is able to load it correctly any longer.