Better error message when replacing an assoc fails
When replacing a has-one association and the replaced record fails to
save (e.g. because a validator triggers), the correct error should be

    RecordNotSaved, "Failed to save the new associated #{reflection.name}."

However, set_owner_attributes fails with the much less descriptive error

    Can't modify frozen hash

if `remove_target!` was called earlier:

    remove_target!(options[:dependent]) if target && !target.destroyed? && assigning_another_record

This solution will skip the failing code, skipping to the correct error
message.

However, it skips set_owner_attributes completely, so the previous state
will not be restored - which is at least not worse than what happened in
the "Can't modify frozen hash" case.