Move belongs_to's `association_primary_key` into `BelongsToReflection`
Most association key methods on the base `AssociationReflection` class
are for has_many and has_one associations, but only
`association_primary_key` is for belongs_to association, it is very
confusing, and so it had a bug for has_one association.

This assorts association key methods on `AssociationReflection`, to have
methods for has_many and has_one associations, and move belongs_to's one
into `BelongsToReflection`.