Enroll new apps in decrypted diffs of credentials
Users, especially those who are creating an app for the first time,
might not be aware of the `rails credentials:diff --enroll` command.
This commit enrolls new apps in decrypted diffs by default so that users
benefit automatically.  This behavior can be opted out of with the app
generator's `--skip-decrypted-diffs` flag.