Add tests for translate() within a partial block
As a follow-up to [PR#39730][] (which proposes this exact diff to the
`6-0-stable` branch) this commit adds test coverage for calling
`translate()` from within the block of a partial.

There are no implementation changes necessary to pass these tests, but
it's possible that the underlying fix for the existing bug was
unintentionally introduced in [dd7a673][] (since neither [PR#39060][]
nor the [CHANGELOG][] mention the fix explicitly).

Adding test coverage at this point would commit to supporting the new
behavior, and would guard against accidental regressions in the future.

[PR#39730]: https://github.com/rails/rails/pull/39730
[dd7a673]: https://github.com/rails/rails/commit/dd7a673782f102d2b58194007bf18a8e496818e3#diff-cda406008ce3935156e8ba173b1e9837
[PR#39060]: https://github.com/rails/rails/pull/39060
[CHANGELOG]: https://github.com/rails/rails/blob/357e15290ab06af8b4f32dd480b48a8c6e31e9e7/actionview/CHANGELOG.md