Add Active Record Marshal forward compatibility tests
This is following a recent discussion on Basecamp.

So far Rails haven't tried to preserve Marshal compatibilty of Active Record instances,
and it probably shouldn't as it isn't as controllable as other serialization formats.

However it's common to store ActiveRecord instances in Rails.cache, and when the
Marshal format changes, it causes problems during Rails upgrade. Process using the new Rails
version might not be able to deserialize the existing cache, and also older process that
weren't restarted yet my fail to load the cache entries written by newer processes.

The goal of this PR is not to freeze the Marshal format, but to have a better visibility
when it changes, so that's it's done on purpose with good reasons rather than accidentally.

Note that this only test forward compatibility (AR vN reading entries written by AR vN-1),
ideally we'd also test backward compatibility (AR vN-1 reading entries written by AR vN),
but it means installing and loading and older Active Record as part of the test suite.