Pass default parameters to url helper
Prior to this commit, if we had a route like this:

    scope "(:account)" do
        resources :courses
    end

and if we were on example.com/5/courses, calling courses_path would generate example.com/courses and it would forget about the optional scope.

In this commit we are passing the default parameters from the URL to the url helper, so that we do not need to pass the scope parameter manually every time we use the helper.