Update association callback docs
There is an old open issue (https://github.com/rails/rails/issues/14365)
saying that callbacks are not triggered when assocations are destroyed.

I believe the issue does not represent a bug, but maybe just a small
gap in the docs about triggering callbacks when removing associations.

This update changes the example of the section to use the `Firm` class.
so that we can include an `after_remove`. This also adds a small `Note:`
section reminding users on how to destory associations so that callbacks
are triggered.