Raise if camelcase inflection impossible in generator
When passing a camelcased model name to a generator sometimes it's
impossible to inflect the casing.  For example "BFF" will generate a
route for the controller BffsController, but the controller generated is
BfFsController.

The casing is impossible to inflect if both:
- a camelcased model name is passed to a generator
- `name.pluralize.underscore.singularize != name.underscore.singularize`

If this happens we raise an error. The error message explains the user
can either us an underscored name or add an inflection instead.