Use headings to make Action Cable terminology more clear [skip ci]
The Terminology section of the Action Cable guide currently is one large
text without any headings. Instead of having one large text, we can use
headings to describe all terminology for easier scanability and improved
table of contents.

The Broadcastings section mentions Pub/Sub so it's moved beneath the Pub/Sub
section.

The following part is moved up to the Connections terminology section.
It makes more sense to have it where it introduces the term.
> Connections form the foundation of the client-server relationship.