Fix highlighting for mostly SQL block
When using Rouge's Console lexer it expects all ignored line breaks to
be proceeded by a backslash. In this code block as we are switching to
Mysql's command line which doesn't need a backslash. This leads to the
lexer breaking and all the lines ending up in a confusing mess.

However the sql syntax actually works much better for this block
anyway.

This line here shows why what we were trying to do wouldn't work with
Rouge:
https://github.com/rouge-ruby/rouge/blob/844c0eeddf7626605fa04b901df6c15c953811d6/lib/rouge/lexers/console.rb#L106