Retain original attribute type when decorating
`AttributeDecorators` was removed in #39882 in favor of calling the
`attribute` macro with a block.  However, this technique overwrote the
original attribute type, which is still needed by some decorators.  Thus
those decorators had to bypass the encapsulation of the `Attributes`
module to retrieve the original attribute type.

This commit makes the `attribute` macro retain the original attribute
type when a new type is not specified.