Skip foreign key rename table test on MariaDB
This test first started to fail when MariaDB 10.4.13 was released.
To fix the build, the `supports_rename_index?` guard was added in
185f935b188606108a7b19817e58153fd81f424d, which meant that the test
was never run on MariaDB since the stable release was still 10.4.

When MariaDB 10.5.4 was released and 10.5 became the stable release, the
test was no longer skipped and started to fail again. It didn't fail on
master because RENAME COLUMN is supported there, which doesn't trigger
the problem.

It looks like the failure is a consequence of fixing this issue:
https://jira.mariadb.org/browse/MDEV-22465